--- tests/queries/0_stateless/02674_space_function.reference
@@ -1,25 +1,0 @@
-   
-   3
-  2
- 1
-0
-
-            12
-          10
-    4
-     5
-
-    4
-                     21
-         9
-       7
-
-                                                       56
-                    20
-     5
-       7
-
-NULL
-NULL
-NULL
-NULL
