+++ tests/queries/0_stateless/02675_space_function.reference
@@ -1,0 +1,21 @@
+   
+   	3
+  	2
+ 	1
+	0
+            	12
+          	10
+    	4
+     	5
+    	4
+                     	21
+         	9
+       	7
+                                                        	56
+                    	20
+     	5
+       	7
+\N
+\N
+\N
+\N
