+++ tests/queries/0_stateless/02675_space_function.sql
